chore(dbnexus): bump 0.3.2 + examples publish=false + workspace member
- Bump version 0.3.1 -> 0.3.2
- Add trait-kit as workspace member with path dependency
- Configure examples package as publish=false
- Add required-features=["kit"] to exclude broken DbNexusKit examples
- Fix clippy lints: manual_range_contains, default_constructed_unit_structs,
  to_string_in_format_args, manual_contains, unnecessary_cast,
  unused_imports (ActiveModelTrait), manual_div_ceil
- Add #![allow(dead_code)] to examples common module
- Add CC0-1.0 to deny.toml licenses allow list (tiny-keccak transitive dep)
- Update Cargo.lock: crossbeam-epoch 0.9.18 -> 0.9.20 (RUSTSEC-2026-0204)
- Add repository metadata to CLI Cargo.toml
- Update README with workspace notes and version
- Bump trait-kit dependency constraint to >=0.2.3

+++ examples/src/observability/metrics_prometheus.rs
@@ -57,7 +57,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     for i in 1..=20 {
         let latency_ms = if i % 5 == 0 { 120 } else { 5 + (i % 10) };
         let success = i != 15; // 第 15 次模拟失败
-        let bytes = if success { Some(100 * i as u64) } else { None };
+        let bytes = if success { Some(100 * i) } else { None };
         collector.record_query("SELECT", Duration::from_millis(latency_ms), success, bytes);
     }
     // 记录少量 INSERT/UPDATE 用于多类型指标展示
